整理搜索页代码
