+++ src/BiliLite.UWP/Extensions/ViewModelExtensions.cs
@@ -5,6 +5,7 @@ using BiliLite.ViewModels.Download;
 using BiliLite.ViewModels.Home;
 using BiliLite.ViewModels.Live;
 using BiliLite.ViewModels.Rank;
+using BiliLite.ViewModels.Search;
 using BiliLite.ViewModels.Settings;
 using BiliLite.ViewModels.User;
 using BiliLite.ViewModels.User.SendDynamic;
@@ -53,6 +54,7 @@ namespace BiliLite.Extensions
             services.AddTransient<ShortcutKeySettingsControlViewModel>();
             services.AddTransient<DevSettingsControlViewModel>();
             services.AddTransient<MainPageViewModel>();
+            services.AddTransient<SearchPageViewModel>();
             return services;
         }
     }
