+++ src/BiliLite.UWP/Models/Common/Enumerates.cs
@@ -525,4 +525,40 @@
         User,
         Desc,
     }
+
+    public enum SearchType
+    {
+        /// <summary>
+        /// 视频
+        /// </summary>
+        Video = 0,
+        /// <summary>
+        /// 番剧
+        /// </summary>
+        Anime = 1,
+        /// <summary>
+        /// 直播
+        /// </summary>
+        Live = 2,
+        /// <summary>
+        /// 主播
+        /// </summary>
+        Anchor = 3,
+        /// <summary>
+        /// 用户
+        /// </summary>
+        User = 4,
+        /// <summary>
+        /// 影视
+        /// </summary>
+        Movie = 5,
+        /// <summary>
+        /// 专栏
+        /// </summary>
+        Article = 6,
+        /// <summary>
+        /// 话题
+        /// </summary>
+        Topic = 7
+    }
 }
