+++ src/BiliLite.UWP/Models/Common/Search/SearchFilterItem.cs
@@ -1,0 +1,13 @@
+﻿namespace BiliLite.Models.Common.Search
+{
+    public class SearchFilterItem
+    {
+        public SearchFilterItem(string name, string value)
+        {
+            this.name = name;
+            this.value = value;
+        }
+        public string name { get; set; }
+        public string value { get; set; }
+    }
+}
