+++ src/BiliLite.UWP/Models/Common/Search/SearchParameter.cs
@@ -1,0 +1,9 @@
+﻿namespace BiliLite.Models.Common.Search
+{
+    public class SearchParameter
+    {
+        public string Keyword { get; set; }
+
+        public SearchType SearchType { get; set; } = SearchType.Video;
+    }
+}
