+++ src/BiliLite.UWP/Models/Common/Search/SearchUserOfficialVerifyItem.cs
@@ -1,0 +1,8 @@
+﻿namespace BiliLite.Models.Common.Search
+{
+    public class SearchUserOfficialVerifyItem
+    {
+        public string desc { get; set; }
+        public int type { get; set; }
+    }
+}
