+++ src/BiliLite.UWP/Startup.cs
@@ -1,5 +1,6 @@
 ﻿using BiliLite.Extensions;
 using BiliLite.Services;
+using BiliLite.Services.Biz;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
@@ -28,6 +29,7 @@ namespace BiliLite
             services.AddBizServices();
             services.AddSingleton<PlaySpeedMenuService>();
             services.AddSingleton<ContentFilterService>();
+            services.AddSingleton<SearchService>();
 
             services.AddSingleton<GrpcService>();
         }
