+++ src/BiliLite.UWP/ViewModels/Search/SearchTopicViewModel.cs
@@ -1,0 +1,83 @@
+﻿using BiliLite.Models;
+using Newtonsoft.Json.Linq;
+using Newtonsoft.Json;
+using System;
+using System.Collections.ObjectModel;
+using System.Threading.Tasks;
+using BiliLite.Extensions;
+using BiliLite.Models.Common.Search;
+
+namespace BiliLite.ViewModels.Search
+{
+    public class SearchTopicViewModel : BaseSearchPivotViewModel
+    {
+        public ObservableCollection<SearchTopicItem> Topics { get; set; }
+
+        public override async Task LoadData()
+        {
+            try
+            {
+                if (Loading)
+                {
+                    return;
+                }
+                ShowLoadMore = false;
+                Loading = true;
+                Nothing = false;
+                var results = await SearchApi.WebSearchTopic(Keyword, Page, Area).Request();
+                if (results.status)
+                {
+                    var data = await results.GetJson<ApiDataModel<JObject>>();
+                    if (data.success)
+                    {
+
+                        var result = JsonConvert.DeserializeObject<ObservableCollection<SearchTopicItem>>(data.data["result"]?.ToString() ?? "[]");
+                        if (Page == 1)
+                        {
+                            if (result == null || result.Count == 0)
+                            {
+                                Nothing = true;
+                                Topics?.Clear();
+                                return;
+                            }
+                            Topics = result;
+                        }
+                        else
+                        {
+                            if (data.data != null)
+                            {
+                                foreach (var item in result)
+                                {
+                                    Topics.Add(item);
+                                }
+                            }
+                        }
+                        if (Page < data.data["numPages"].ToInt32())
+                        {
+                            ShowLoadMore = true;
+                            Page++;
+                        }
+                        HasData = true;
+                    }
+                    else
+                    {
+                        Notify.ShowMessageToast(data.message);
+                    }
+                }
+                else
+                {
+                    Notify.ShowMessageToast(results.message);
+                }
+            }
+            catch (Exception ex)
+            {
+                var handel = HandelError<SearchPageViewModel>(ex);
+                Notify.ShowMessageToast(handel.message);
+            }
+            finally
+            {
+                Loading = false;
+            }
+        }
+    }
+}
